refactor: ContentView.swift内で定義されていたSafariViewを切り出し、SafariView.swiftを作成
